x/fdpd/keeper/grpc_query_decision_policy.go
package keeper

import (
	"context"

	"delegationcontrol/x/fdpd/types"
	sdk "github.com/cosmos/cosmos-sdk/types"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"
)

func (k Keeper) DecisionPolicy(c context.Context, req *types.QueryGetDecisionPolicyRequest) (*types.QueryGetDecisionPolicyResponse, error) {
	if req == nil {
		return nil, status.Error(codes.InvalidArgument, "invalid request")
	}
	ctx := sdk.UnwrapSDKContext(c)

	val, found := k.GetDecisionPolicy(ctx)
	if !found {
		return nil, status.Error(codes.NotFound, "not found")
	}

	return &types.QueryGetDecisionPolicyResponse{DecisionPolicy: val}, nil
}
